feat(library): Introduce base (ELF Loader) to library
Introduce `base` image (ELF Loader) for bincompat runs.

Add `Kraftfile` to build for QMEU and Firecracker on x86_64 using initrd
for volume support. It uses the `testing` branch of Unikraft, that
incorporates all not-yet-upstream changes.

Add GitHub workflow to build and deploy the image. Add success badge in
`README.md`.

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>